services/backend/requirements.txt
fastapi
psycopg[binary]
pydantic-settings
strawberry-graphql[fastapi]
uvicorn[standard]

